# Improved and batched logs translation