--- a/internal/stanza/factory_test.go
+++ b/internal/stanza/factory_test.go
@@ -17,6 +17,7 @@ package stanza
 import (
 	"context"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/component"
@@ -35,6 +36,18 @@ func TestCreateReceiver(t *testing.T) {
 			{
 				"type": "json_parser",
 			},
+		}
+		receiver, err := factory.CreateLogsReceiver(context.Background(), params, cfg, &mockLogsConsumer{})
+		require.NoError(t, err, "receiver creation failed")
+		require.NotNil(t, receiver, "receiver creation failed")
+	})
+
+	t.Run("Success with ConverterConfig", func(t *testing.T) {
+		factory := NewFactory(TestReceiverType{})
+		cfg := factory.CreateDefaultConfig().(*TestConfig)
+		cfg.Converter = ConverterConfig{
+			MaxFlushCount: 1,
+			FlushInterval: 3 * time.Second,
 		}
 		receiver, err := factory.CreateLogsReceiver(context.Background(), params, cfg, &mockLogsConsumer{})
 		require.NoError(t, err, "receiver creation failed")