--- a/receiver/syslogreceiver/testdata/config.yaml
+++ b/receiver/syslogreceiver/testdata/config.yaml
@@ -3,6 +3,8 @@ receivers:
     tcp:
       listen_address: "0.0.0.0:29018"
     protocol: rfc5424
+    converter:
+      flush_interval: 100ms
 
 processors:
   nop: