… receiver/filelogreceiver/filelog_test.go
… receiver/filelogreceiver/go.sum
… receiver/syslogreceiver/go.sum