@tigrannajaryan 2021-03-17
@pmalek-sumo @djaglowski I haven't looked into this PR details so my comment may be wrong, but if I understand correctly this PR attempts to group logs after they are produced by github.com/open-telemetry/opentelemetry-log-collection

Can we instead change how they are produced? Presumably we collect a batch or log lines at once from each file that we tail. Can we send the batch as it is collected from the file instead of sending one log line at a time and then trying to batch later?

The benefits of the suggested approach would be:
- No extra delays introduced due to the new batching logic.
- Hopefully less performance impact.