@djaglowski 2021-03-17
@tigrannajaryan I think your suggestion may be a good idea to explore, but it would be a fairly substantial change.

> Presumably we collect a batch or log lines at once from each file that we tail.

This presumption is not accurate. Logs "lines" are processed one at a time, with the exception being for multiline logs which are treated as one log entry.

When it comes down to it, most operations supported by opentelemetry-log-collection must be performed on each log entry individually, and each subsequent operation is basically just a function on a call stack, so batching doesn't provide a lot of additional performance in many cases. 

You certainly have a good point in regards to the value that would be provided by keeping related entries together to avoid having to reaggregate them later. We can probably do this, but there are a lot of implications to think through and I don't believe it is a good next step.

@pmalek-sumo's change does appear to have performance benefits in it's own right, so I suggest we should try to adopt the changes at this level and remove them later if we can make them redundant.