@pmalek-sumo 2021-03-17
> Can we send the batch as it is collected from the file instead of sending one log line at a time and then trying to batch later?

That could be an improvement (taking into account @djaglowski comments though) but this PR tries to batch logs per resource which as I understand the concepts used in OTC doesn't translate 1:1 into a tailed file, correct?

Hence the batching done with this approach is a bit different than what you're proposing @tigrannajaryan. Please correct me if I'm wrong.