@tigrannajaryan 2021-03-17
> This presumption is not accurate. Logs "lines" are processed one at a time, with the exception being for multiline logs which are treated as one log entry.

Got it.

>Hence the batching done with this approach is a bit different than what you're proposing @tigrannajaryan. Please correct me if I'm wrong.

Correct. I was suggesting that a batch log lines that is read from a file at once is split, parsed and delivered by log-collection library as a slice of entry.LogEntry at once, instead of delivering them one by one.

Such a slice of LogEntrys would still be one pdata.Logs and would be emitted by the receiver using one ConsumeLogs call. However, you are right that these LogEntrys do not necessarily have the same Resource, so that's a different way to group them.

Grouping by the Resource is a different approach and it is more expensive. I would want to know what the performance impact of this approach is. The serialization of the Resource, lookup in the map have costs, I am curious how big it is. Alternatively, we can simply group subsequent LogEntrys which have the same Resource, without reordering and grouping. That's less expensive to do and may yield similar final results when reading at high speed from a small set of files.