@pmalek-sumo 2021-03-17
One more thing: what is the benefit of receiving the logs grouped by Resource downstream? Any particular optimizations in terms of performance can be/are done in downstream consumers in such a scenario?