@djaglowski 2021-03-19
@pmalek-sumo Any idea why CI is hanging on this PR?

Since the main purpose of this PR is to increase performance, can you please post before/after results for related testbed scenarios?