@pmalek-sumo 2021-03-19
Ok I guess I've figured it out.

---

baseline: https://github.com/open-telemetry/opentelemetry-collector-contrib/commit/3bdf190c0aee4eb1928ca7a9a3de311f1d4772a6

```
# Test PerformanceResults
Started: Fri, 19 Mar 2021 15:08:19 +0100

Test                                    |Result|Duration|CPU Avg%|CPU Max%|RAM Avg MiB|RAM Max MiB|Sent Items|Received Items|
----------------------------------------|------|-------:|-------:|-------:|----------:|----------:|---------:|-------------:|
Log10kDPS/OTLP                          |PASS  |     17s|    18.6|    16.3|         38|         49|    150000|        150000|
Log10kDPS/filelog                       |PASS  |     15s|    22.5|    23.0|         41|         49|    150000|        150000|
Log10kDPS/kubernetes_containers         |PASS  |     15s|    49.1|    50.3|         45|         54|    150000|        150000|
Log10kDPS/k8s_CRI-Containerd            |PASS  |     15s|    44.5|    44.7|         46|         55|    150000|        150000|
Log10kDPS/k8s_CRI-Containerd_no_attr_ops|PASS  |     15s|    39.6|    40.7|         45|         53|    150000|        150000|
Log10kDPS/CRI-Containerd                |PASS  |     15s|    24.6|    25.0|         44|         52|    150000|        150000|

Total duration: 94s
```

---

this PR at https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/2694/commits/e59915f2ac2b62cc0393f342904b6d3bb9cac7c0

first run:

```
# Test PerformanceResults
Started: Fri, 19 Mar 2021 14:49:30 +0100

Test                                    |Result|Duration|CPU Avg%|CPU Max%|RAM Avg MiB|RAM Max MiB|Sent Items|Received Items|
----------------------------------------|------|-------:|-------:|-------:|----------:|----------:|---------:|-------------:|
Log10kDPS/OTLP                          |PASS  |     17s|    17.3|    14.7|         39|         49|    149900|        149900|
Log10kDPS/filelog                       |PASS  |     15s|    17.6|    20.3|         40|         49|    150000|        150000|
Log10kDPS/kubernetes_containers         |PASS  |     15s|    46.5|    47.0|         44|         53|    150000|        150000|
Log10kDPS/k8s_CRI-Containerd            |PASS  |     15s|    41.7|    42.6|         44|         54|    150000|        150000|
Log10kDPS/k8s_CRI-Containerd_no_attr_ops|PASS  |     15s|    34.0|    34.3|         44|         54|    150000|        150000|
Log10kDPS/CRI-Containerd                |PASS  |     15s|    20.1|    20.7|         42|         51|    150000|        150000|

Total duration: 94s
```

second run:

```
# Test PerformanceResults
Started: Fri, 19 Mar 2021 15:14:46 +0100

Test                                    |Result|Duration|CPU Avg%|CPU Max%|RAM Avg MiB|RAM Max MiB|Sent Items|Received Items|
----------------------------------------|------|-------:|-------:|-------:|----------:|----------:|---------:|-------------:|
Log10kDPS/OTLP                          |PASS  |     17s|    19.5|    18.3|         40|         49|    149900|        149900|
Log10kDPS/filelog                       |PASS  |     15s|    18.6|    19.9|         40|         49|    150000|        150000|
Log10kDPS/kubernetes_containers         |PASS  |     15s|    47.5|    49.1|         44|         54|    150000|        150000|
Log10kDPS/k8s_CRI-Containerd            |PASS  |     15s|    42.2|    43.0|         45|         54|    150000|        150000|
Log10kDPS/k8s_CRI-Containerd_no_attr_ops|PASS  |     15s|    34.8|    36.0|         43|         52|    150000|        150000|
Log10kDPS/CRI-Containerd                |PASS  |     15s|    21.7|    22.6|         42|         51|    150000|        150000|

Total duration: 94s
```